My_Captain_task_4:- JAVA program for matrix addition and multiplication of two matrices
This program in JAVA takes number of rows and columns of each matrices as an input from user and then takes the value of each elements in the matrix from the user. The user has to make sure that THE NUMBER OF ROWS AND COLUMNS IN THE FIRST MATRIX EQUALS SECOND. If THE NUMBER OF ROWS AND COLUMNS IN THE FIRST MATRIX EQUALS SECOND ,the program will give an output of addition and multiplication of the matrices, else it will show an error.